refactor: converted test to typescript
